add game init after both players present
